LayoutTests/platform/ios-simulator/editing/selection/caret-in-empty-inline-1-expected.txt

Bug 85793: Caret is not rendered in empty inline contenteditable elements

This test verifies that an empty inline contenteditable element gets a valid caret rect.


PASS caretRect.left is 8
PASS caretRect.top is 160
FAIL caretRect.width should be 1. Was 2.
PASS caretRect.height is 20
PASS successfullyParsed is true

TEST COMPLETE
